refact: move competence goals to article
