+++ src/components/Article/CompetenceGoalsDialog.js
