+++ src/components/Article/index.js
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2018-present, NDLA.
+ *
+ * This source code is licensed under the GPLv3 license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+
+import Article from './Article';
+
+export default Article;
